Move custom configuration cache reporting logic to buildSrc
