+++ buildSrc/subprojects/uber-plugins/src/main/kotlin/gradlebuild.internal.kotlin-js.gradle.kts
@@ -42,7 +42,11 @@ kotlin {
 
 tasks {
     withType<KotlinJsCompile>().configureEach {
-        kotlinOptions.allWarningsAsErrors = true
+        kotlinOptions {
+            allWarningsAsErrors = true
+            metaInfo = false
+            moduleKind = "plain"
+        }
     }
 
     val ktlintCheckTasks = withType<KtlintCheckTask>()
